backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
dnspython==2.4.2

# AI and Vector database for RAG
openai==1.6.1
qdrant-client==1.7.0
cohere==4.38

# Redis for caching and rate limiting
redis==5.0.1
hiredis==2.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Development tools
python-dotenv==1.0.0
